remove housekeeping for active trades

+++ service/housekeeper.py
@@ -1,6 +1,8 @@
 import model
 import helper
 import asyncio
+from service.trades import Trades
+from service.data import Data
 from datetime import datetime, timedelta
 
 logging = helper.LoggerFactory.get_logger("logs/housekeeper.log", "housekeeper")
@@ -21,12 +23,17 @@ class Housekeeper:
                 hours=self.housekeeping_interval
             )
             try:
-                query = await model.Tickers.filter(
-                    timestamp__lt=cleanup_timestamp.timestamp()
-                ).delete()
-                logging.info(
-                    f"Start housekeeping. Delete {query} entries older then {cleanup_timestamp}"
-                )
+                active_symbols = Trades().get_symbols()
+                ticker_symbols = Data().get_ticker_symbol_list()
+                # Do not housekeep active trades
+                for symbol in ticker_symbols:
+                    if symbol not in active_symbols:
+                        query = await model.Tickers.filter(
+                            timestamp__lt=cleanup_timestamp.timestamp()
+                        ).delete()
+                        logging.info(
+                            f"Start housekeeping. Delete {query} entries older then {cleanup_timestamp}"
+                        )
             except Exception as e:
                 logging.error(f"Error db housekeeping: {e}")
